Move network related module from jormungandr-testing-utils to hersir
