+++ testing/hersir/Cargo.toml
@@ -26,6 +26,10 @@ serde_json = "1.0"
 structopt = "0.3.25"
 thiserror = "1.0"
 yaml-rust = "0.4.4"
+rand = "0.8"
+rand_core = "0.6"
+rand_chacha = "0.3"
+multiaddr = { package = "parity-multiaddr", version = "0.11" }
 
 [dependencies.reqwest]
 version = "0.11"
