+++ testing/hersir/src/builder/vote_plan_settings.rs
@@ -1,6 +1,6 @@
-use crate::wallet::PrivateVoteCommitteeDataManager;
 use chain_impl_mockchain::certificate::VotePlanId;
 use jormungandr_lib::interfaces::VotePlan;
+use jormungandr_testing_utils::wallet::PrivateVoteCommitteeDataManager;
 
 #[derive(Debug)]
 pub enum VotePlanSettings {
