+++ testing/hersir/src/builder/wallet/template/builder.rs
@@ -1,9 +1,8 @@
-use crate::testing::network::{
-    wallet::{WalletAlias, WalletTemplate},
-    NodeAlias,
-};
+use crate::builder::WalletTemplate;
 use chain_addr::Discrimination;
 use chain_impl_mockchain::value::Value;
+use jormungandr_testing_utils::testing::node::NodeAlias;
+use jormungandr_testing_utils::wallet::WalletAlias;
 
 pub struct WalletTemplateBuilder {
     alias: WalletAlias,
