+++ testing/hersir/src/builder/wallet/template/mod.rs
@@ -1,12 +1,13 @@
 pub mod builder;
 mod external;
 
-use super::{WalletAlias, WalletType};
-use crate::testing::network::NodeAlias;
+use super::WalletType;
 use chain_addr::Discrimination;
 use chain_impl_mockchain::value::Value;
 pub use external::ExternalWalletTemplate;
 use jormungandr_lib::interfaces::{DiscriminationDef, ValueDef};
+use jormungandr_testing_utils::testing::node::NodeAlias;
+use jormungandr_testing_utils::wallet::WalletAlias;
 use serde::Deserialize;
 
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Hash)]
