+++ testing/hersir/src/controller/interactive/mod.rs
@@ -2,9 +2,9 @@ pub mod args;
 mod command;
 mod controller;
 
+use crate::controller::Error as ControllerError;
 pub use command::InteractiveCommand;
 pub use controller::{do_for_all_alias, UserInteractionController};
-
 pub use jortestkit::prelude::{ConsoleWriter, InteractiveCommandError, InteractiveCommandExec};
 use std::ffi::OsStr;
 use structopt::StructOpt;
