+++ testing/hersir/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod args;
+pub mod builder;
 pub mod cli;
 pub mod config;
 pub mod controller;
