+++ testing/jormungandr-testing-utils/src/testing/jormungandr/mod.rs
@@ -2,7 +2,7 @@ mod configuration_builder;
 pub mod process;
 pub mod starter;
 
-use crate::testing::network::NodeAlias;
+use crate::testing::node::NodeAlias;
 use crate::testing::{FragmentNode, FragmentNodeError, MemPoolCheck};
 use chain_core::property::Fragment as _;
 use chain_impl_mockchain::fragment::Fragment;
